+++ views/layouts/main.php
@@ -13,7 +13,7 @@ use app\core\Application;
     <link rel="apple-touch-icon" sizes="76x76" href="../assets/img/apple-icon.png">
     <link rel="icon" type="image/png" href="../assets/img/favicon.png">
     <title>
-        VBIS 2024/2025
+        CCC
     </title>
     <!--     Fonts and icons     -->
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700" rel="stylesheet"/>
@@ -24,82 +24,82 @@ use app\core\Application;
     <script src="https://kit.fontawesome.com/42d5adcbca.js" crossorigin="anonymous"></script>
     <!-- CSS Files -->
     <link id="pagestyle" href="../assets/css/argon-dashboard.css?v=2.1.0" rel="stylesheet"/>
-
     <link rel="stylesheet" href="../assets/js/plugins/toastr/toastr.min.css">
     <script src="../assets/libs/jquery/dist/jquery.min.js"></script>
     <script src="../assets/js/plugins/toastr/toastr.min.js"></script>
     <script src="../assets/js/plugins/toastr/toastr-options.js"></script>
 </head>
 
-<body class="g-sidenav-show   bg-gray-100">
-<div class="min-height-300 bg-dark position-absolute w-100"></div>
-<aside class="sidenav bg-white navbar navbar-vertical navbar-expand-xs border-0 border-radius-xl my-3 fixed-start ms-4 "
-       id="sidenav-main">
-    <div class="sidenav-header">
-        <i class="fas fa-times p-3 cursor-pointer text-secondary opacity-5 position-absolute end-0 top-0 d-none d-xl-none"
-           aria-hidden="true" id="iconSidenav"></i>
-        <a class="navbar-brand m-0" href="/"
-           target="_blank">
-            <img src="../assets/img/logo-ct-dark.png" width="26px" height="26px" class="navbar-brand-img h-100"
-                 alt="main_logo">
-            <span class="ms-1 font-weight-bold"> VBIS 2024/2025</span>
-        </a>
-    </div>
-    <hr class="horizontal dark mt-0">
-    <div class="collapse navbar-collapse  w-auto " id="sidenav-collapse-main">
-        <ul class="navbar-nav">
-            <li class="nav-item">
-                <a class="nav-link " href="/">
-                    <div class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
-                        <i class="ni ni-tv-2 text-dark text-sm opacity-10"></i>
+<body class="">
+<div class="container position-sticky z-index-sticky top-0">
+    <div class="row">
+        <div class="col-12">
+            <!-- Navbar -->
+            <nav class="navbar navbar-expand-lg blur border-radius-lg top-0 z-index-3 shadow position-absolute mt-4 py-2 start-0 end-0 mx-4">
+                <div class="container-fluid">
+                    <a class="navbar-brand font-weight-bolder ms-lg-0 ms-3 " href="/login">
+                        CCC
+                    </a>
+                    <button class="navbar-toggler shadow-none ms-2" type="button" data-bs-toggle="collapse"
+                            data-bs-target="#navigation" aria-controls="navigation" aria-expanded="false"
+                            aria-label="Toggle navigation">
+              <span class="navbar-toggler-icon mt-2">
+                <span class="navbar-toggler-bar bar1"></span>
+                <span class="navbar-toggler-bar bar2"></span>
+                <span class="navbar-toggler-bar bar3"></span>
+              </span>
+                    </button>
+                    <div class="collapse navbar-collapse justify-content-end" id="navigation">
+                        <ul class="navbar-nav">
+                            <?php
+                            if (Application::$app->session->get('user')) {
+                                echo '<li class="nav-item me-0">';
+                                echo '<a class="nav-link me-2" href="/processLogout">';
+                                echo '<i class="ni ni-user-run opacity-6 text-dark me-1"></i>';
+                                echo 'Logout';
+                                echo '</a>';
+                                echo '</li>';
+                            } else {
+                                echo '<li class="nav-item me-0">';
+                                echo '<a class="nav-link me-2" href="/login">';
+                                echo '<i class="fas fa-key opacity-6 text-dark me-1"></i>';
+                                echo ' Sign In';
+                                echo '</a>';
+                                echo '</li>';
+                            }
+                            ?>
+                        </ul>
                     </div>
-                    <span class="nav-link-text ms-1">Home</span>
-                </a>
-            </li>
-            <li class="nav-item mt-3">
-                <h6 class="ps-4 ms-2 text-uppercase text-xs font-weight-bolder opacity-6">Account pages</h6>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link " href="/users">
-                    <div class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
-                        <i class="ni ni-single-02 text-dark text-sm opacity-10"></i>
-                    </div>
-                    <span class="nav-link-text ms-1">Users</span>
-                </a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link " href="/products">
-                    <div class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
-                        <i class="ni ni-folder-17 text-dark text-sm opacity-10"></i>
-                    </div>
-                    <span class="nav-link-text ms-1">Products</span>
-                </a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link " href="/processLogout">
-                    <div class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
-                        <i class="ni ni-user-run text-dark text-sm opacity-10"></i>
-                    </div>
-                    <span class="nav-link-text ms-1">Log out</span>
-                </a>
-            </li>
-        </ul>
-    </div>
-    <div class="sidenav-footer mx-3 ">
-        <div class="card card-plain shadow-none" id="sidenavCard">
-            <img class="w-50 mx-auto" src="../assets/img/illustrations/icon-documentation.svg"
-                 alt="sidebar_illustration">
-            <div class="card-body text-center p-3 w-100 pt-0">
-            </div>
+                </div>
+            </nav>
+            <!-- End Navbar -->
         </div>
     </div>
-</aside>
-<main class="main-content position-relative border-radius-lg ">
-    <div class="container-fluid py-4">
-        {{ RENDER_SECTION }}
-    </div>
+</div>
+<main class="main-content  mt-0">
+    <section>
+        <div class="page-header min-vh-100">
+            <div class="container">
+                <div class="row">
+                    <div class="col-xl-4 col-lg-5 col-md-7 d-flex flex-column mx-lg-0 mx-auto">
+                        {{ RENDER_SECTION }}
+                    </div>
+                    <div class="col-6 d-lg-flex d-none h-100 my-auto pe-0 position-absolute top-0 end-0 text-center justify-content-center flex-column">
+                        <div class="position-relative bg-gradient-primary h-100 m-3 px-7 border-radius-lg d-flex flex-column justify-content-center overflow-hidden"
+                             style="background-image: url('https://raw.githubusercontent.com/creativetimofficial/public-assets/master/argon-dashboard-pro/assets/img/signin-ill.jpg');
+          background-size: cover;">
+                            <span class="mask bg-gradient-primary opacity-6"></span>
+                            <h4 class="mt-5 text-white font-weight-bolder position-relative">"Attention is the new
+                                currency"</h4>
+                            <p class="text-white position-relative">The more effortless the writing looks, the more
+                                effort the writer actually put into the process.</p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
 </main>
-
 <!--   Core JS Files   -->
 <script src="../assets/js/core/popper.min.js"></script>
 <script src="../assets/js/core/bootstrap.min.js"></script>
@@ -124,5 +124,4 @@ use app\core\Application;
 Application::$app->session->showSuccessNotification();
 Application::$app->session->showErrorNotification();
 ?>
-
 </html>
